# make the queue fully nogc and optimize it with a freelist