Since this queue is implemented as a singly linked list, the nodes that are popped can be directly used as a free list.