@Herringway 2018-06-12
This isn't really my area of expertise, but I'm interested in seeing some benchmarks nonetheless