@BBasile 2018-06-12
For this benchmark:

```d
void main()
{
    import std.datetime.stopwatch;
    import std.stdio;

    void test()
    {
        auto queue = Queue!int();
        foreach(i; 0 .. 1000000)
        {
            foreach(j; 0 .. 100) queue.push(i); // after: for each i > 0 we alloc 90 and recycle 10
            foreach(j; 0 .. 10) queue.pop;
        }
    }
    benchmark!test(1)[0].writeln;
}
```

- **before**: 8 secs, 616 ms, 333 μs, and 2 hnsecs
- **after**: 6 secs, 847 ms, 188 μs, and 1 hnsec

Now and TBH, i dont know if this is a realistic scenario in D:YAML.