@BBasile 2018-06-12
Actually this should be tested with a D-YAML type... `int` does not lead to add a GC range.