@BBasile 2018-06-13
Last commit allow ~6% gain on a big 50Mo yaml file.