@BBasile 2018-06-13
I think it's quite safe to merge this ?

- module is dedicated to this queue ( i.e easily reversible)
- test suite passes
- there will be beta for 0.7.0 where it will tested in actual projects.